worker/requirements.txt
flask
pymupdf
pillow
pytesseract
zxing-cpp
flask-cors
